moved prompt func to output file
