+++ cmd/utils/output.go
@@ -3,6 +3,7 @@ package utils
 import (
 	"errors"
 	"fmt"
+	"github.com/manifoldco/promptui"
 	"math/big"
 	"os"
 	"time"
@@ -59,6 +60,21 @@ func GetLoadingSpinner() *spinner.Spinner {
 type OutputHandler struct {
 	NoOutput bool
 	spinner  *spinner.Spinner
+}
+
+func PromptBool(msg string) (bool, error) {
+	prompt := promptui.Prompt{
+		Label:     msg,
+		IsConfirm: true,
+	}
+	_, err := prompt.Run()
+	if err != nil {
+		if err == promptui.ErrAbort {
+			return false, nil
+		}
+		return false, err
+	}
+	return true, nil
 }
 
 func NewOutputHandler(noOutput bool) *OutputHandler {
