+++ utils/filesystem.go
@@ -2,7 +2,6 @@ package utils
 
 import (
 	"fmt"
-	"github.com/manifoldco/promptui"
 	"os"
 )
 
@@ -21,19 +20,4 @@ func DirNotEmpty(path string) (bool, error) {
 
 	files, err := os.ReadDir(path)
 	return len(files) > 0, err
-}
-
-func PromptBool(msg string) (bool, error) {
-	prompt := promptui.Prompt{
-		Label:     msg,
-		IsConfirm: true,
-	}
-	_, err := prompt.Run()
-	if err != nil {
-		if err == promptui.ErrAbort {
-			return false, nil
-		}
-		return false, err
-	}
-	return true, nil
 }
